If there are custom fonts, default select the Filter as that is the common case.
Switched font service to using the user signal.

Properly prompt to ensure force deletion is handled.

Handle shifting selected font and filter when deleting.

+++ API/Controllers/FontController.cs
@@ -81,16 +81,30 @@ public class FontController : BaseApiController
     /// Removes a font from the system
     /// </summary>
     /// <param name="fontId"></param>
-    /// <param name="confirmed">If the font is in use by other users and an admin wants it deleted, they must confirm to force delete it</param>
+    /// <param name="force">If the font is in use by other users and an admin wants it deleted, they must confirm to force delete it. This is prompted in the UI.</param>
     /// <returns></returns>
     [HttpDelete]
-    public async Task<IActionResult> DeleteFont(int fontId, bool confirmed = false)
+    public async Task<IActionResult> DeleteFont(int fontId, bool force = false)
     {
-        // TODO: We need to check if this font is used by anyone else and if so, need to inform the user
-        // Need to check if this is a system font as well
-        var forceDelete = User.IsInRole(PolicyConstants.AdminRole) && confirmed;
-        await _fontService.Delete(fontId);
+        var forceDelete = User.IsInRole(PolicyConstants.AdminRole) && force;
+        var fontInUse = await _fontService.IsFontInUse(fontId);
+        if (!fontInUse || forceDelete)
+        {
+            await _fontService.Delete(fontId);
+        }
+
         return Ok();
+    }
+
+    /// <summary>
+    /// Returns if the given font is in use by any other user. System provided fonts will always return true.
+    /// </summary>
+    /// <param name="fontId"></param>
+    /// <returns></returns>
+    [HttpGet("in-use")]
+    public async Task<ActionResult<bool>> IsFontInUse(int fontId)
+    {
+        return Ok(await _fontService.IsFontInUse(fontId));
     }
 
     /// <summary>
